+++ objects/GXDLMSRegisterMonitor.go
@@ -143,7 +143,7 @@ func (g *GXDLMSRegisterMonitor) GetValue(settings *settings.GXDLMSSettings, e *i
 		if err != nil {
 			return nil, err
 		}
-		err = internal.SetData(settings, data, enums.DataTypeUint16, g.MonitoredValue.ObjectType)
+		err = internal.SetData(settings, data, enums.DataTypeUint16, uint16(g.MonitoredValue.ObjectType))
 		if err != nil {
 			return nil, err
 		}
@@ -497,7 +497,7 @@ func (g *GXDLMSRegisterMonitor) GetDataType(index int) (enums.DataType, error) {
 // NewGXDLMSRegisterMonitor creates a new register monitor object instance.
 //
 // The function validates `ln` before creating the object.
-//`ln` is the Logical Name and `sn` is the Short Name of the object.
+// `ln` is the Logical Name and `sn` is the Short Name of the object.
 func NewGXDLMSRegisterMonitor(ln string, sn int16) (*GXDLMSRegisterMonitor, error) {
 	err := ValidateLogicalName(ln)
 	if err != nil {
